+++ src/commands/command/codecs/update-time-to-live.ts
@@ -1,0 +1,7 @@
+import { encode } from "../inputs/update-time-to-live.js";
+import { decode } from "../outputs/update-time-to-live.js";
+
+export const UpdateTimeToLiveCodec = {
+  encode,
+  decode,
+};