+++ src/commands/command/outputs/update-time-to-live.ts
@@ -1,0 +1,9 @@
+import type { UpdateTimeToLiveCommandOutput } from "@aws-sdk/client-dynamodb";
+
+export type UpdateTimeToLiveOutput = Record<string, never>;
+
+export function decode(
+  _: UpdateTimeToLiveCommandOutput,
+): UpdateTimeToLiveOutput {
+  return {};
+}